Refactor EventMenu and UserEventListItem components to streamline event actions
* Consolidated event action handlers into a new useEventActions hook for better modularity and reusability.
* Updated EventMenu to utilize the new hook, simplifying the component and improving readability.
* Introduced UserEventListItem component with integrated event action handlers for sharing, directions, and calendar additions.
* Enhanced icon usage and layout for a more consistent user experience across event-related actions.

These changes improve code maintainability and user interaction within the Expo app.
